% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPTraceR.R
\docType{package}
\name{CPTrackR}
\alias{CPTrackR}
\title{CPTrackR:  Create unique track identifiers for CellProfiler tracking output}
\description{
By default CellProfiler does not export unique track ids, this package fixes that.
}
