% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPTraceR.R
\name{createLUT}
\alias{createLUT}
\title{Fix the track ids for multiple group}
\usage{
createLUT(data)
}
\arguments{
\item{data}{The dataframe with groupNumber, groupInd, par_obj_num, obj_num}
}
\description{
Optionally in parallel. (doesn't handle plate ID or other column names yet)
}
