% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPTraceR.R
\name{createLUTGroup}
\alias{createLUTGroup}
\title{Fix the track ids for a single group (="movie")}
\usage{
createLUTGroup(group_df)
}
\arguments{
\item{group_df}{The dataframe with data for a single group}
}
\description{
Fix the track ids for a single group (="movie")
}
